MLS counter buff
Yesterday I removed the accuracy bonus on MLS, for a couple reasons. First of all I think the tighter spread made it difficult to even do damage at all for players who aren't exactly the best at aiming, but it also made it overpowered in the hands of a skilled player.

This nerf might have helped players actually hit targets, but it made mid range damage (the range where most players aimed just fine at) pathetic, seems like the nerf may have been over-done. But instead of restoring the accuracy, I just increased the damage a tad with this commit.

If this ends up being too powerful, what I will most likely do is reduce the clip size to 3, to match what it is in PLR. This would give the weapon a bit less staying power when it comes to dealing with swarms of opponents, and it would also reduce the amount of burst damage the giant could put out against Engineer buildings, which to me could be a good thing, as MLS has ironically been one of the best giants for destroying sentries.